refactor: organizar lógica de usuario y categoría separando diseño de la lógica en formularios e índices
