+++ frontend/src/app/dashboard/access/users/page.tsx
@@ -1,0 +1,5 @@
+import UsersPage from "@/features/dashboard/Users";
+
+export default function Page() {
+  return <UsersPage />;
+}
