+++ frontend/src/features/dashboard/CategoryProducts/hooks/useCategories.ts
@@ -1,0 +1,78 @@
+import { useState } from "react";
+import { Category, CreateCategoryData } from "../types";
+
+// Datos mock iniciales
+const initialCategories: Category[] = [
+  { id: 1, nombre: "Electrónicos", descripcion: "Dispositivos electrónicos y gadgets", estado: "Activo" },
+  { id: 2, nombre: "Ropa", descripcion: "Prendas de vestir para todas las edades", estado: "Activo" },
+  { id: 3, nombre: "Hogar", descripcion: "Artículos para el hogar y decoración", estado: "Inactivo" },
+  { id: 4, nombre: "Deportes", descripcion: "Equipamiento y ropa deportiva", estado: "Activo" },
+  { id: 5, nombre: "Juguetes", descripcion: "Juguetes para niños y niñas", estado: "Activo" },
+  { id: 6, nombre: "Libros", descripcion: "Libros de todos los géneros", estado: "Activo" },
+  { id: 7, nombre: "Belleza", descripcion: "Productos de cuidado personal y belleza", estado: "Inactivo" },
+  { id: 8, nombre: "Alimentos", descripcion: "Productos alimenticios y bebidas", estado: "Activo" },
+  { id: 9, nombre: "Muebles", descripcion: "Muebles para interior y exterior", estado: "Activo" },
+  { id: 10, nombre: "Jardín", descripcion: "Herramientas y plantas para jardín", estado: "Activo" },
+  { id: 11, nombre: "Tecnología", descripcion: "Dispositivos tecnológicos y accesorios", estado: "Activo" },
+  { id: 12, nombre: "Salud", descripcion: "Productos para el cuidado de la salud", estado: "Inactivo" },
+  { id: 13, nombre: "Automóviles", descripcion: "Accesorios y repuestos para autos", estado: "Activo" },
+  { id: 14, nombre: "Instrumentos", descripcion: "Instrumentos musicales y accesorios", estado: "Activo" },
+  { id: 15, nombre: "Oficina", descripcion: "Suministros y muebles de oficina", estado: "Activo" },
+  { id: 16, nombre: "Bebés", descripcion: "Productos para bebés y niños pequeños", estado: "Inactivo" },
+  { id: 17, nombre: "Mascotas", descripcion: "Alimentos y accesorios para mascotas", estado: "Activo" },
+  { id: 18, nombre: "Viajes", descripcion: "Equipaje y accesorios de viaje", estado: "Activo" },
+  { id: 19, nombre: "Joyeria", descripcion: "Joyas y accesorios personales", estado: "Activo" },
+  { id: 20, nombre: "Herramientas", descripcion: "Herramientas manuales y eléctricas", estado: "Inactivo" },
+  { id: 21, nombre: "Arte", descripcion: "Materiales y suministros de arte", estado: "Activo" },
+  { id: 22, nombre: "Videojuegos", descripcion: "Consolas y videojuegos", estado: "Activo" },
+  { id: 23, nombre: "Fotografía", descripcion: "Cámaras y equipos de fotografía", estado: "Activo" },
+  { id: 24, nombre: "Relojes", descripcion: "Relojes de pulsera y de pared", estado: "Inactivo" },
+  { id: 25, nombre: "Calzado", descripcion: "Zapatos and calzado para toda ocasión", estado: "Activo" },
+  { id: 26, nombre: "Outdoor", descripcion: "Equipamiento para actividades al aire libre", estado: "Activo" },
+  { id: 27, nombre: "Limpieza", descripcion: "Productos de limpieza para el hogar", estado: "Activo" },
+  { id: 28, nombre: "Electrodomésticos", descripcion: "Electrodomésticos grandes y pequeños", estado: "Inactivo" },
+  { id: 29, nombre: "Navidad", descripcion: "Decoraciones y artículos navideños", estado: "Activo" },
+  { id: 30, nombre: "Coleccionables", descripcion: "Artículos de colección y antigüedades", estado: "Activo" },
+];
+
+export const useCategories = () => {
+  const [categories, setCategories] = useState<Category[]>(initialCategories);
+  const [isCreateModalOpen, setIsCreateModalOpen] = useState(false);
+
+  const handleCreateCategory = (categoryData: CreateCategoryData) => {
+    const newCategory: Category = {
+      id: categories.length + 1,
+      nombre: categoryData.nombre,
+      descripcion: categoryData.descripcion,
+      estado: "Activo",
+    };
+
+    setCategories(prev => [...prev, newCategory]);
+    setIsCreateModalOpen(false);
+  };
+
+  const handleView = (category: Category) => {
+    console.log("Ver categoría:", category);
+    // Aquí puedes implementar la lógica para ver la categoría
+  };
+
+  const handleEdit = (category: Category) => {
+    console.log("Editar categoría:", category);
+    // Aquí puedes implementar la lógica para editar la categoría
+  };
+
+  const handleDelete = (category: Category) => {
+    console.log("Eliminar categoría:", category);
+    // Aquí puedes implementar la lógica para eliminar la categoría
+  };
+
+  return {
+    categories,
+    isCreateModalOpen,
+    setIsCreateModalOpen,
+    handleCreateCategory,
+    handleView,
+    handleEdit,
+    handleDelete
+  };
+};
