+++ frontend/src/features/dashboard/CategoryProducts/types/index.ts
@@ -1,0 +1,12 @@
+export interface Category {
+  id: number;
+  nombre: string;
+  descripcion: string;
+  estado: "Activo" | "Inactivo";
+}
+
+export interface CreateCategoryData {
+  nombre: string;
+  descripcion: string;
+  icono: File | null;
+}
