+++ frontend/src/features/dashboard/Purchases.tsx
@@ -1,7 +1,8 @@
 "use client";
 
+import RequireAuth from "../auth/requireauth";
 import { Column, DataTable } from "./components/DataTable";
-import RequireAuth from "../auth/requireAuth";
+
 
 // Tipo de datos de cada compra
 type Purchase = {
